+++ docs/history.rst
@@ -5,6 +5,7 @@ Version History
 --------------------------------
 - **Bump minimum Python version to 3.9**
 - Add a transcoder for `application/x-www-formurlencoded`_
+- Add support for encoding :class:`decimal.Decimal`
 - Add type annotations (see :ref:`type-info`)
 - Return a "406 Not Acceptable" if the :http:header:`Accept` header values cannot be matched
   and there is no default content type configured